% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solocp_single.R
\name{solocp_single}
\alias{solocp_single}
\title{solocp_single compute the marginal inclusion probabilities for all time instance in the case n_t=1}
\usage{
solocp_single(
  y,
  sigma2,
  q = 1,
  tau2 = NULL,
  tau2.spike = NULL,
  tau2.slab = NULL
)
}
\arguments{
\item{y}{vector of observations}

\item{sigma2}{sample variance}

\item{q}{sparsity inducing parameters (if not included set to 0.1)}

\item{tau2}{shrinkage parameter (if not included set to 2/sqrt(n))}

\item{tau2.spike}{spike variance parameters (if not included set to 1/n)}

\item{tau2.slab}{slab variance parameters (if not included set to n)}
}
\value{
Weight of the spike compenent  \code{w.spike}
Weight of the slab compenent  \code{w.slab}
inclusion probability \code{ratio}
}
\description{
solocp_single compute the marginal inclusion probabilities for all time instance in the case n_t=1
}
