% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_changepoints.R
\name{subset_changepoints}
\alias{subset_changepoints}
\title{Subset the change points computed either basad.cp or solo.cp
Corresponds to point 2-4 in Algorithm 1 in the manuscript}
\usage{
subset_changepoints(ratio, del = 5)
}
\arguments{
\item{ratio}{marginal inclusion probabilities, as computed for example by the solocp_single function}

\item{del}{Delta parameter in the parameter, mininum spacing between change points to be considered nonconsecutive
(we set it equal to 5 by defauly)}
}
\value{
subset of change points indexes
}
\description{
Subset the change points computed either basad.cp or solo.cp
Corresponds to point 2-4 in Algorithm 1 in the manuscript
}
